GITHUB-APD-143: Re-organize Bundle to add backend
