+++ composer.json
@@ -14,7 +14,8 @@
         },
         "psr-4": {
             "Pim\\Upgrade\\": "upgrades/",
-            "Akeneo\\Connectivity\\Connection\\": "src/Akeneo/Connectivity/Connection/back"
+            "Akeneo\\Connectivity\\Connection\\": "src/Akeneo/Connectivity/Connection/back",
+            "Akeneo\\Platform\\CommunicationChannel\\": "src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back"
         }
     },
     "autoload-dev": {
@@ -29,7 +30,8 @@
             "AkeneoTest\\": "tests/back",
             "Akeneo\\Test\\IntegrationTestsBundle\\": "tests/back/Integration/IntegrationTestsBundle/",
             "Akeneo\\Test\\": "tests/back/",
-            "Pim\\Upgrade\\Schema\\Tests\\": "upgrades/test_schema/"
+            "Pim\\Upgrade\\Schema\\Tests\\": "upgrades/test_schema/",
+            "Akeneo\\Platform\\CommunicationChannel\\Test\\": "src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/tests"
         }
     },
     "require": {
