+++ package.json
@@ -136,6 +136,6 @@
     "src/Akeneo/Connectivity/Connection/front",
     "src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/legacy-bridge",
     "src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared",
-    "src/Akeneo/Platform/Bundle/CommunicationChannelBundle/Resources/workspaces/communication-channel"
+    "src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front"
   ]
 }
