+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/Infrastructure/Symfony/Resources/public/images/announcements/error-monitoring-connection-dashboard.png
