+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/Infrastructure/Symfony/Resources/public/images/announcements/number-of-products-sent.png
