+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/Infrastructure/Symfony/Resources/public/less/components/Menu.less
