+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/Infrastructure/Symfony/Resources/public/menu/communication-channel.ts
