+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/Infrastructure/Symfony/Resources/translations/jsmessages.en_US.yml
