+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/.eslintrc.json
