+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "main": "src/index.ts",
   "config": {
-    "generate_models_script": "./../../../../../../../../frontend/build/json-schema-to-typescript.js"
+    "generate_models_script": "./../../../../../../frontend/build/json-schema-to-typescript.js"
   },
   "scripts": {
     "generate-models": "node $npm_package_config_generate_models_script src/definitions/ src/models/"
