+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/icons/Megaphone.tsx
