+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/index.tsx
