+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/panel/Header.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import {AkeneoThemedProps, CloseIcon, useAkeneoTheme} from '@akeneo-pim-community/shared';
 import {useTranslate} from '@akeneo-pim-community/legacy-bridge';
-import {MegaphoneIcon} from './../icons/Megaphone';
+import {MegaphoneIcon} from '../icons/Megaphone';
 
 const Container = styled.div`
   margin-right: 30px;
