+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/panel/announcement/Tag.tsx
