+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/panel/index.ts
