+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/definitions/announcement.schema.json
