+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/definitions/pimAnalyticsData.schema.json
