+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/hooks/useAnnouncements.ts
