+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/hooks/usePimVersion.ts
