+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/index.ts
