+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/models/.gitignore
