+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/tools/formatCampaign.ts
