+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/validator/pimAnalyticsData.ts
