+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/tests/front/unit/__mocks__/dataProvider.ts
