+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/tests/front/unit/hooks/usePimVersion.unit.ts
