+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/tests/front/unit/tools/formatCampaign.unit.ts
