+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/tsconfig.json
